requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapy==1.0.5
h11==0.16.0
httptools==0.6.4
idna==3.10
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
fastapi
uvicorn[standard]
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
pydantic
pyodbc
psycopg2-binary
